+++ src/audio/kpb.c
@@ -29,6 +29,7 @@
 #include <sof/math/numbers.h>
 #include <sof/platform.h>
 #include <sof/schedule/schedule.h>
+#include <sof/schedule/task.h>
 #include <sof/string.h>
 #include <sof/ut.h>
 #include <ipc/topology.h>
