+++ src/drivers/dw/ssi-spi.c
@@ -19,6 +19,7 @@
 #include <sof/lib/wait.h>
 #include <sof/platform.h>
 #include <sof/schedule/schedule.h>
+#include <sof/schedule/task.h>
 #include <sof/spinlock.h>
 #include <sof/string.h>
 #include <ipc/header.h>
