+++ src/include/sof/drivers/ipc.h
@@ -11,7 +11,7 @@
 
 #include <sof/list.h>
 #include <sof/platform.h>
-#include <sof/schedule/schedule.h>
+#include <sof/schedule/task.h>
 #include <sof/spinlock.h>
 #include <sof/trace/dma-trace.h>
 #include <sof/trace/trace.h>
