+++ src/include/sof/schedule/edf_schedule.h
@@ -8,6 +8,7 @@
 #ifndef __SOF_SCHEDULE_EDF_SCHEDULE_H__
 #define __SOF_SCHEDULE_EDF_SCHEDULE_H__
 
+#include <sof/schedule/task.h>
 #include <sof/trace/trace.h>
 #include <stdint.h>
