+++ src/include/sof/schedule/schedule.h
@@ -34,25 +34,6 @@ enum {
 	SOF_SCHEDULE_COUNT
 };
 
-#define SOF_TASK_PRI_HIGH	0	/* priority level 0 - high */
-#define SOF_TASK_PRI_MED	4	/* priority level 4 - medium */
-#define SOF_TASK_PRI_LOW	9	/* priority level 9 - low */
-
-#define SOF_TASK_PRI_COUNT	10	/* range of priorities (0-9) */
-
-#define SOF_TASK_PRI_IPC	SOF_TASK_PRI_LOW
-#define SOF_TASK_PRI_IDC	SOF_TASK_PRI_LOW
-
-/* task states */
-#define SOF_TASK_STATE_INIT		0
-#define SOF_TASK_STATE_QUEUED		1
-#define SOF_TASK_STATE_PENDING		2
-#define SOF_TASK_STATE_RUNNING		3
-#define SOF_TASK_STATE_PREEMPTED	4
-#define SOF_TASK_STATE_COMPLETED	5
-#define SOF_TASK_STATE_FREE		6
-#define SOF_TASK_STATE_CANCEL		7
-
 /* Scheduler flags */
 /* Sync/Async only supported by ll scheduler atm */
 #define SOF_SCHEDULE_FLAG_ASYNC (0 << 0) /* task scheduled asynchronously */
@@ -71,20 +52,6 @@ struct scheduler_ops {
 	int (*scheduler_init)(void);
 	void (*scheduler_free)(void);
 	void (*scheduler_run)(void);
-};
-
-struct task {
-	uint16_t type;
-	uint64_t start;
-	uint16_t priority;
-	uint16_t state;
-	uint16_t core;
-	void *data;
-	uint64_t (*func)(void *data);
-	struct list_item list;
-	struct list_item irq_list;	/* list for assigned irq level */
-	const struct scheduler_ops *ops;
-	void *private;
 };
 
 struct schedule_data {
