+++ src/lib/agent.c
@@ -19,6 +19,7 @@
 #include <sof/debug/panic.h>
 #include <sof/platform.h>
 #include <sof/schedule/schedule.h>
+#include <sof/schedule/task.h>
 #include <sof/sof.h>
 #include <sof/trace/trace.h>
 #include <ipc/topology.h>
