+++ src/schedule/ll_schedule.c
@@ -33,6 +33,7 @@
 #include <sof/platform.h>
 #include <sof/schedule/ll_schedule.h>
 #include <sof/schedule/schedule.h>
+#include <sof/schedule/task.h>
 #include <sof/spinlock.h>
 #include <ipc/topology.h>
 #include <errno.h>
