+++ test/cmocka/src/audio/kpb/kpb_buffer.c
@@ -23,6 +23,7 @@
 #include "kpb_mock.h"
 #include <sof/list.h>
 #include <sof/lib/notifier.h>
+#include <sof/schedule/task.h>
 #include <user/kpb.h>
 
 /*! Local data types */
